refactor: split argument structs into their own module for easier readability
