+++ src/lookup/types.rs
@@ -1,5 +1,6 @@
 use crate::get_name;
-use crate::utils::cli::{self, TypeArgs};
+use crate::utils::args::TypeArgs;
+use crate::utils::cli;
 use crate::utils::helpers;
 use clap::error::ErrorKind;
 use rustemon::Follow;
