+++ src/utils.rs
@@ -1,3 +1,4 @@
+pub mod args;
 pub mod cli;
 pub mod enums;
 pub mod helpers;
