+++ src/utils/cli.rs
@@ -1,6 +1,6 @@
-use crate::utils::enums::*;
+use crate::utils::args;
 use clap::builder::styling::{AnsiColor, Effects, Style, Styles};
-use clap::{Args, CommandFactory, Parser, Subcommand};
+use clap::{CommandFactory, Parser, Subcommand};
 
 pub const HEADER: Style = AnsiColor::Green.on_default().effects(Effects::BOLD);
 pub const USAGE: Style = AnsiColor::Green.on_default().effects(Effects::BOLD);
@@ -46,11 +46,11 @@ pub struct Cli {
 pub enum Command {
   /// Look up the varieties of a given pokemon.
   #[command(name = "list", long_about)]
-  ListCmd(ListArgs),
+  ListCmd(args::ListArgs),
 
   /// Look up the type(s) of a given pokemon.
   #[command(name = "types", long_about)]
-  TypeCmd(TypeArgs),
+  TypeCmd(args::TypeArgs),
 
   /// Look up the abilities of a given pokemon. If the ability is a hidden ability, it will be
   /// marked accordingly.
@@ -59,7 +59,7 @@ pub enum Command {
     about = "Look up the abilities of a given pokemon",
     long_about
   )]
-  AbilityCmd(AbilityArgs),
+  AbilityCmd(args::AbilityArgs),
 
   /// Look up the level-up moveset of a given pokemon. If a level is provided
   /// then the four moves at or below the given level are provided. By default, this will
@@ -69,327 +69,32 @@ pub enum Command {
     about = "Look up the level-up moveset of a given pokemon",
     long_about
   )]
-  MoveCmd(MoveArgs),
+  MoveCmd(args::MoveArgs),
 
   /// Look up the egg groups of a given pokemon species.
   #[command(name = "eggs", long_about)]
-  EggCmd(EggArgs),
+  EggCmd(args::EggArgs),
 
   /// Look up the gender ratio of a given pokemon species.
   #[command(name = "genders", long_about)]
-  GenderCmd(GenderArgs),
+  GenderCmd(args::GenderArgs),
 
   /// Look up the encounters for a given pokemon and version.
   #[command(name = "encounters", long_about)]
-  EncounterCmd(EncounterArgs),
+  EncounterCmd(args::EncounterArgs),
 
   /// Look up evolution chain for a given pokemon species.
   #[command(name = "evolutions", long_about)]
-  EvolutionCmd(EvolutionArgs),
+  EvolutionCmd(args::EvolutionArgs),
 
   /// Look up the type weaknesses/resistances for given type(s).
   #[command(name = "matchups", long_about)]
-  MatchupCmd(MatchupArgs),
+  MatchupCmd(args::MatchupArgs),
 
   /// Open web pages for a given endpoint. A valid endpoint includes pokemon, abilities, items, and more.
   #[cfg(feature = "web")]
   #[command(name = "web", long_about)]
-  WebCmd(WebArgs),
-}
-
-// ========================================================================================================================
-// Subcommand Arguments
-// ========================================================================================================================
-
-#[derive(Args, Debug)]
-pub struct ListArgs {
-  #[arg(help = "name of pokemon species")]
-  pub pokemon: String,
-
-  #[arg(short, long, help = "skip API requests for formatted names")]
-  pub fast: bool,
-
-  #[arg(value_enum,
-    short = 'L',
-    long,
-    value_name = "LANGUAGE",
-    default_value_t = LanguageId::En,
-    hide_possible_values=true,
-    help = "language ID for API requests for formatted names"
-  )]
-  pub lang: LanguageId,
-}
-
-#[derive(Args, Debug)]
-pub struct TypeArgs {
-  #[arg(help = "name of pokemon")]
-  pub pokemon: String,
-
-  #[arg(short, long, help = "skip API requests for formatted names")]
-  pub fast: bool,
-
-  #[arg(short, long = "gen", help = "generation to query")]
-  pub generation: Option<i64>,
-
-  #[arg(value_enum,
-    short = 'L',
-    long,
-    value_name = "LANGUAGE",
-    default_value_t = LanguageId::En,
-    hide_possible_values=true,
-    help = "language ID for API requests for formatted names"
-  )]
-  pub lang: LanguageId,
-
-  #[arg(short, help = "recursively check evolution chain")]
-  pub recursive: bool,
-}
-
-#[derive(Args, Debug)]
-pub struct AbilityArgs {
-  #[arg(help = "name of pokemon")]
-  pub pokemon: String,
-
-  #[arg(short, long, help = "skip API requests for formatted names")]
-  pub fast: bool,
-
-  #[arg(value_enum,
-    short = 'L',
-    long,
-    value_name = "LANGUAGE",
-    default_value_t = LanguageId::En,
-    hide_possible_values=true,
-    help = "language ID for API requests for formatted names"
-  )]
-  pub lang: LanguageId,
-
-  #[arg(short, help = "recursively check evolution chain")]
-  pub recursive: bool,
-}
-
-#[derive(Args, Debug)]
-pub struct MoveArgs {
-  #[arg(help = "name of pokemon")]
-  pub pokemon: String,
-
-  #[arg(short, long, help = "skip API requests for formatted names")]
-  pub fast: bool,
-
-  #[arg(value_enum,
-    short = 'L',
-    long,
-    value_name = "LANGUAGE",
-    default_value_t = LanguageId::En,
-    hide_possible_values=true,
-    help = "language ID for API requests for formatted names"
-  )]
-  pub lang: LanguageId,
-
-  #[arg(value_enum, short, long, default_value_t=VersionGroup::ScarletViolet,
-          hide_possible_values=true, help="version group name")]
-  pub vgroup: VersionGroup,
-
-  #[arg(short, long, help = "request default moveset at given level")]
-  pub level: Option<i64>,
-}
-
-#[derive(Args, Debug)]
-pub struct EggArgs {
-  #[arg(help = "name of pokemon species")]
-  pub pokemon: String,
-
-  #[arg(short, long, help = "skip API requests for formatted names")]
-  pub fast: bool,
-
-  #[arg(value_enum,
-    short = 'L',
-    long,
-    value_name = "LANGUAGE",
-    default_value_t = LanguageId::En,
-    hide_possible_values=true,
-    help = "language ID for API requests for formatted names"
-  )]
-  pub lang: LanguageId,
-}
-
-#[derive(Args, Debug)]
-pub struct GenderArgs {
-  #[arg(help = "name of pokemon species")]
-  pub pokemon: String,
-
-  #[arg(short, long, help = "skip API requests for formatted names")]
-  pub fast: bool,
-
-  #[arg(value_enum,
-    short = 'L',
-    long,
-    value_name = "LANGUAGE",
-    default_value_t = LanguageId::En,
-    hide_possible_values=true,
-    help = "language ID for API requests for formatted names"
-  )]
-  pub lang: LanguageId,
-}
-
-#[derive(Args, Debug)]
-pub struct EncounterArgs {
-  #[arg(value_enum, hide_possible_values = true, help = "name of version")]
-  pub version: Version,
-
-  #[arg(help = "name of pokemon")]
-  pub pokemon: String,
-
-  #[arg(short, long, help = "skip API requests for formatted names")]
-  pub fast: bool,
-
-  #[arg(value_enum,
-    short = 'L',
-    long,
-    value_name = "LANGUAGE",
-    default_value_t = LanguageId::En,
-    hide_possible_values=true,
-    help = "language ID for API requests for formatted names"
-  )]
-  pub lang: LanguageId,
-
-  #[arg(short, help = "recursively check evolution chain")]
-  pub recursive: bool,
-
-  #[arg(short, long, help = "condense output by only showing location areas")]
-  pub condensed: bool,
-}
-
-#[derive(Args, Debug)]
-pub struct EvolutionArgs {
-  #[arg(help = "name of pokemon species")]
-  pub pokemon: String,
-
-  #[arg(short, long, help = "skip API requests for formatted names")]
-  pub fast: bool,
-
-  #[arg(value_enum,
-    short = 'L',
-    long,
-    value_name = "LANGUAGE",
-    default_value_t = LanguageId::En,
-    hide_possible_values=true,
-    help = "language ID for API requests for formatted names"
-  )]
-  pub lang: LanguageId,
-
-  #[arg(
-    short,
-    long,
-    help = "hide the names of the pokemon in the evolution chain"
-  )]
-  pub secret: bool,
-
-  #[arg(short, long, help = "show all evolution chains, even outdated ones")]
-  pub all: bool,
-}
-
-#[derive(Args, Debug)]
-pub struct MatchupArgs {
-  #[arg(
-    value_enum,
-    hide_possible_values = true,
-    value_name = "TYPE",
-    help = "name of type"
-  )]
-  pub primary: Type,
-
-  #[arg(
-    value_enum,
-    hide_possible_values = true,
-    value_name = "TYPE",
-    help = "name of optional secondary type"
-  )]
-  pub secondary: Option<Type>,
-
-  #[arg(short, long, help = "print output as a list instead of a table")]
-  pub list: bool,
-
-  #[arg(short, long, help = "skip API requests for formatted names")]
-  pub fast: bool,
-
-  #[arg(value_enum,
-    short = 'L',
-    long,
-    value_name = "LANGUAGE",
-    default_value_t = LanguageId::En,
-    hide_possible_values=true,
-    help = "language ID for API requests for formatted names"
-  )]
-  pub lang: LanguageId,
-}
-
-#[derive(Args, Debug)]
-pub struct WebArgs {
-  #[command(flatten)]
-  pub endpoint: Endpoints,
-
-  #[arg(short = 'A', long, help = "name of area within region")]
-  pub area: Option<String>,
-
-  #[arg(short, long = "gen", help = "optional name of generation to use")]
-  pub generation: Option<i64>,
-
-  #[arg(short, long, help = "suppress print statements")]
-  pub quiet: bool,
-}
-
-#[cfg(feature = "web")]
-#[derive(Args, Debug)]
-#[group(required = true, multiple = false)]
-pub struct Endpoints {
-  #[arg(short, long, help_heading = "Endpoints", conflicts_with_all = ["area"], help = "name of pokemon")]
-  pub pokemon: Option<String>,
-
-  #[arg(short, long, help_heading = "Endpoints", help = "name of region")]
-  pub region: Option<String>,
-
-  #[arg(
-    short,
-    long,
-    help_heading = "Endpoints",
-    conflicts_with = "area",
-    help = "name of move"
-  )]
-  pub move_: Option<String>,
-
-  #[arg(short, long, help_heading = "Endpoints", conflicts_with_all = ["area", "generation"], help = "name of ability")]
-  pub ability: Option<String>,
-
-  #[arg(short, long, help_heading = "Endpoints", conflicts_with_all = ["area", "generation"], help = "name of item")]
-  pub item: Option<String>,
-}
-
-#[cfg(feature = "web")]
-impl Endpoints {
-  pub fn get_mode(&self) -> DexMode {
-    if let Some(name) = &self.pokemon {
-      DexMode::Pokedex(name.clone())
-    } else if let Some(name) = &self.region {
-      DexMode::Pokearth(name.clone())
-    } else if let Some(name) = &self.move_ {
-      DexMode::Attackdex(name.clone())
-    } else if let Some(name) = &self.ability {
-      DexMode::Abilitydex(name.clone())
-    } else if let Some(name) = &self.item {
-      DexMode::Itemdex(name.clone())
-    } else {
-      unreachable!()
-    }
-  }
-}
-
-#[cfg(feature = "web")]
-pub enum DexMode {
-  Pokedex(String),
-  Pokearth(String),
-  Attackdex(String),
-  Abilitydex(String),
-  Itemdex(String),
+  WebCmd(args::WebArgs),
 }
 
 // ========================================================================================================================
